refactor: ManageControllerを責務ごとに分割
- ManageController: ビュー表示のみ（index, show, settings）
- ManageChannelApiController: チャンネルAPI（fetchChannel, addChannel）
- ManageArchiveApiController: アーカイブAPI（fetchArchives, addArchives, toggleDisplay, fetchComments, editTimestamps）
- ManageSettingsApiController: 設定API（除外ワード, カバー曲）
- ManageAccessControl trait: canAccessChannel共通ロジック
- MappingStatusHelper: マッピング状態判定の共通ヘルパー

ルート名は変更なし。

Closes #488
